Update V.2.0.11
The title and icon of the website was modified, navigation was added to the contact button in the Main.

Added the URL to the project data matrix to link the routes to their destination.

Syntax was corrected in Sidenav because it indicated an error in the console, the styles were being called as "class" and correctly it is "className".

Pending:

- Evaluate if it is necessary to place a footer, to do so describe its content and links.

- Evaluate if we place the two resumes in Spanish and English.